# Conflicts:
#	.gitignore
#	server/ruoyi-admin/src/main/resources/application-druid.yml
#	server/sql/health_data_manager.sql
#	web/src/views/ai/index.vue
#	web/src/views/login/index.vue

+++ server/ruoyi-framework/src/main/java/com/ruoyi/framework/web/service/SysRegisterService.java
@@ -102,6 +102,10 @@ public class SysRegisterService
      */
     public void validateCaptcha(String username, String code, String uuid)
     {
+        if (StringUtils.isEmpty(code) && StringUtils.isEmpty(uuid))
+        {
+            return;
+        }
         String verifyKey = CacheConstants.CAPTCHA_CODE_KEY + StringUtils.nvl(uuid, "");
         String captcha = redisCache.getCacheObject(verifyKey);
         redisCache.deleteObject(verifyKey);
